Port initialization code written, but not tested with a real Elm module.
If you copy or soft link example/site/index.js.dummy to index.js, you can then aim a browser at example/site/index.html, and it will call WebSocketClient.subscribe() in example/site/js/WebSocketClient.js to subscribe to the Elm-to-JS port and test for the existence of the JS-to-Elm port.

The action code in example/site/WebSocketClient.js is now commented out, awaiting conversion to work with the ports.