Make it Webpack compatible by elminating the global variable.
example/site/js/PortFunnel.js now uses the same mechanism that Elm does in its generated '.js' files. Instead of a global variable, it passes 'this' into the top-level function, which sets its 'PortFunnel' property. In a '<script>' tag at the top of an HTML file, this is the same as a global  variable, but it also works with Webpack.